cas-api/application/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cas-ng==4.2.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
idna==3.3
lxml==4.6.4
python-cas==1.6.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.0
urllib3==1.26.7
